+++ packages/adhere-ui-historyback/es/historyback.js
@@ -1,2 +1,2 @@
-export default function(e,r){void 0===r&&(r="/"),"undefined"!=typeof document&&(!("referrer"in document)||""!==document.referrer?window.history.go(-1):e&&r&&e.replace(r||"/"))}
+export default function(o,i){void 0===i&&(i="/"),window.history.length?window.history.back():o.replace(i||"/")}
 //# sourceMappingURL=historyback.js.map