+++ packages/adhere-ui-historyback/src/historyback.ts
@@ -4,15 +4,21 @@
  * @param routePath {string}
  */
 export default (history, routePath = '/') => {
-  if(typeof document === 'undefined') return;
-  
-  if ('referrer' in document) {
-    if (document.referrer !== '') {
-      window.history.go(-1);
-    } else if (history && routePath) {
-      history.replace(routePath || '/');
-    }
+  // if(typeof document === 'undefined') return;
+  //
+  // if ('referrer' in document) {
+  //   if (document.referrer !== '') {
+  //     window.history.go(-1);
+  //   } else if (history && routePath) {
+  //     history.replace(routePath || '/');
+  //   }
+  // } else {
+  //   window.history.go(-1);
+  // }
+
+  if (!!window.history.length) {
+    window.history.back();
   } else {
-    window.history.go(-1);
+    history.replace(routePath || '/');
   }
 };